fix: created test process file and imported tellurium for pytest
